moved to profile-directory, additional story added
